allow downloading workspace usage CSV via API
